SAT Collision now fully working! Removed the detection collision, and minor improvements
The problem with my SAT Collision was that I was not resetting the min and maxes after each iteration. I removed the diagnol collision as it falls into the same pitfalls as AABB collision and it was heurstic. Will re-implent a radius detection algorithm but solely for two circles. I had forgotten to add the geometry shader version to the Resource Manager so I added that in. Game class now cleaned up