Particle System fully intregrated
Particle system is now complete and the color bug turned out to be a bug in the overall program. Basically, some textures become transparent if their color is not all 1.0s. The reason for this I am not sure why. could be the blending mode, shader etc. Added velocity and color variations. I decided not to use multithreading for the particle system as I would have to thread the entire class. Would complicate things as I still want massive control over the particle generator in the main thread.