Camera class added
Camera class was added and the ability to follow an entity around was included as well. Only small problem to fix with that is making the camera centered around the entity itself, right now in a weird position. Also adding an option for making the text * by the view matrix.