Text rotation added and entity transparency allowed
You can now control the opacity of the entites through the w component of the color vector. While you are now able to rotate text the pivot is not correct. Reason being the origin is not where it should be placed as my pos is not actually on the top left of the sprite. Instead it's in the middle which is weird. Need to find a way to set pos correctly